Fix the issue that TSI image was not loaded in case of secure boot
Once secure boot is enabled, TSI should be used and initialized at the
boot stage. OPTEE will not load the TSI image if it finds that TSI has
already been initialized. This results in the failure of TSI_OTP_Read()
because this command relies on the TSI image.
This patch fixes the issue by always loading the TSI image.

Change-Id: I447bba0a59ac33b5448a6f8033478b77d737cba3